requirements.txt
mysql-connector-python
python-dotenv
requests
schedule
selenium
webdriver_manager 